NEAT training
git-svn-id: http://encog-cs.googlecode.com/svn/trunk/encog-core@247 0d3f3ea6-ac51-0410-a1de-edb5cd2a854a

+++ encog-core-cs/Solve/Genetic/Innovation/BasicInnovation.cs
@@ -13,6 +13,6 @@ namespace Encog.Solve.Genetic.Innovation
         /// <summary>
         /// The innovation id.
         /// </summary>
-        public int InnovationID { get; set; }
+        public long InnovationID { get; set; }
     }
 }
